Move and rename various tti-tempo files
